Fixed sending data about unloaded chunks over network

+++ src/main/java/com/github/minecraft_ta/totaldebug/companionApp/chunkGrid/IChunkGridManager.java
@@ -2,9 +2,9 @@ package com.github.minecraft_ta.totaldebug.companionApp.chunkGrid;
 
 public interface IChunkGridManager {
 
-    int UNLOADED_CHUNK = 0;
-    int SPAWN_CHUNK = 1;
-    int LAZY_CHUNK = 2;
+    byte UNLOADED_CHUNK = 0;
+    byte SPAWN_CHUNK = 1;
+    byte LAZY_CHUNK = 2;
 
     void update();
 }
